add SQL Manager
