+++ S303-EscapeRoom/src/main/java/org/example/Modules/CLASESTESTS/GameTEST.java
@@ -1,8 +1,8 @@
-package org.example.Modules.Communicates.CLASESTEST;
+package org.example.Modules.CLASESTESTS;
 
+import org.example.Modules.CLASESTESTS.PlayerTEST;
 import org.example.Modules.Communicates.Certificate;
 import org.example.Modules.Communicates.CommunicateType;
-import org.example.Modules.Communicates.Notification;
 
 import java.util.ArrayList;
