+++ S303-EscapeRoom/src/main/java/org/example/Modules/Communicates/Certificate.java
@@ -1,7 +1,7 @@
 package org.example.Modules.Communicates;
 
-import org.example.Modules.Communicates.CLASESTEST.GameTEST;
-import org.example.Modules.Communicates.CLASESTEST.PlayerTEST;
+import org.example.Modules.CLASESTESTS.GameTEST;
+import org.example.Modules.CLASESTESTS.PlayerTEST;
 
 import java.util.Date;
