+++ S303-EscapeRoom/src/main/java/org/example/Modules/Communicates/CommFactory/CommFactoryInterface.java
@@ -1,6 +1,6 @@
 package org.example.Modules.Communicates.CommFactory;
 
-import org.example.Modules.Communicates.CLASESTEST.PlayerTEST;
+import org.example.Modules.CLASESTESTS.PlayerTEST;
 import org.example.Modules.Communicates.Communicate;
 
 public interface CommFactoryInterface {
