+++ S303-EscapeRoom/src/main/java/org/example/Modules/Communicates/CommFactory/TicketFactory.java
@@ -1,7 +1,7 @@
 package org.example.Modules.Communicates.CommFactory;
 
-import org.example.Modules.Communicates.CLASESTEST.PlayerTEST;
-import org.example.Modules.Communicates.CLASESTEST.SaleTEST;
+import org.example.Modules.CLASESTESTS.PlayerTEST;
+import org.example.Modules.CLASESTESTS.SaleTEST;
 import org.example.Modules.Communicates.Ticket;
 
 public class TicketFactory implements CommFactoryInterface{
